Eric Sunshine made the following changes:
-*- Added new interface iParticlesColorCallback.  This replaces the ugly
    functional-style color callback used by iParticlesObjectState and
    iParticlesFactoryState with a more object-oriented and normalized
    approach.

-*- Cleaned up particle related headers and added some missing
    documentation to ensure that Doxygen will actually produce output for
    the interfaces.

-*- Eliminated several compilation warning in the particles and
    particlesldr plugins.


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@16520 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/particles/persist/standard/particlesldr.cpp
@@ -604,7 +604,7 @@ bool csParticlesFactoryLoader::ParseColorHeat (iDocumentNode *node,
       case XMLTOKEN_BASE_HEAT:
       {
         float base_heat = child->GetContentsValueAsFloat ();
-        state->SetHeatColorMethod (base_heat);
+        state->SetHeatColorMethod ((int)base_heat);
         method_set = true;
         break;
       }
@@ -1186,7 +1186,7 @@ bool csParticlesObjectLoader::ParseColorHeat (iDocumentNode *node,
       case XMLTOKEN_BASE_HEAT:
       {
         float base_heat = child->GetContentsValueAsFloat ();
-        state->SetHeatColorMethod (base_heat);
+        state->SetHeatColorMethod ((int)base_heat);
         method_set = true;
         break;
       }
